django-backends/blog/requirements.txt
asgiref==3.3.0
Django==3.1.3
django-ckeditor==6.0.0
django-cors-headers==3.5.0
django-js-asset==1.2.2
django-mdeditor==0.1.18
djangorestframework==3.12.2
Markdown==3.3.3
markdown2==2.3.10
mutagen==1.45.1
mysqlclient==2.0.1
pyMd2Doc==2.0.2
pytz==2020.4
sqlparse==0.4.1
